GOYO-BE/ec2-anc/anc/requirements.txt
# GOYO AI Server Dependencies

fastapi==0.104.1
uvicorn[standard]==0.24.0

numpy==1.24.3
scipy==1.11.3

paho-mqtt==1.6.1

pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

python-multipart==0.0.6